+++ lib/src/bones_ui.dart
@@ -1,3 +1,3 @@
 class BonesUI {
-  static const String version = '2.1.10';
+  static const String version = '2.1.11';
 }